requirements.txt
gpiozero>=1.5.1
paho-mqtt>=1.5.1
rpi.gpio>=0.7.0
